Many fixups (#1)
* Cleanup warnings

* Decouple bars from their extra space.

I wrote these changes a while ago, but never committed them. I'm not
entirely sure what the status was, but things steem to still work.

* Update to 2018 edition

* Fix rust 2018 idioms

* Perform mostly clippy-inspired cleanups

* Add logging and don't kill threads on errors

* Cleanup unused imports

* Use inotify for brightness bar again

* Fix space issue

The newline charater was being provided by each bar, but it needed to be
provided by `BarWithSep`. This also reduces flicker in the `Stdin` bar.